fix(kiloclaw): add pattern check to validateUserFilePath, extract shared validateOpenclawJson
validateUserFilePath now checks USER_FILTERED_PATTERNS (e.g. .bak., .kilotmp.)
to match the filtering applied in filterFileTree, closing a bypass gap where
crafted paths could read/write filtered files directly via tRPC.

Also extracts duplicated openclaw.json client-side validation into a shared
validateOpenclawJsonForSave utility used by both user and admin editors.
