+++ src/app/admin/components/KiloclawInstances/AdminFileEditor.tsx
@@ -2,10 +2,10 @@
 
 import { useCallback } from 'react';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-import { toast } from 'sonner';
 import { useTRPC } from '@/lib/trpc/utils';
 import { FileEditorShell } from '@/app/(app)/claw/components/FileEditorShell';
 import { FileEditorPane, type FileSaveError } from '@/app/(app)/claw/components/FileEditorPane';
+import { validateOpenclawJsonForSave } from '@/app/(app)/claw/components/validateOpenclawJson';
 
 function AdminFileEditorPaneInner({
   userId,
@@ -42,21 +42,7 @@ function AdminFileEditorPaneInner({
     [writeFileMutation, userId]
   );
 
-  const validateBeforeSave = useCallback((path: string, content: string) => {
-    if (path === 'openclaw.json') {
-      try {
-        const parsed = JSON.parse(content);
-        if (typeof parsed !== 'object' || parsed === null || Array.isArray(parsed)) {
-          toast.error('Config must be a JSON object');
-          return false;
-        }
-      } catch {
-        toast.error('Invalid JSON — fix syntax errors before saving');
-        return false;
-      }
-    }
-    return true;
-  }, []);
+  const validateBeforeSave = useCallback(validateOpenclawJsonForSave, []);
 
   return (
     <FileEditorPane
